requirements.txt
# Core dependencies for Push-to-Write
SpeechRecognition==3.10.4
pyaudio==0.2.14
pynput==1.7.7
pyperclip==1.9.0
python-dotenv==1.0.1
pystray==0.19.5
Pillow==10.4.0
openai-whisper==20231117
torch>=2.0.0
numpy>=1.24.0
pyautogui==0.9.54
plyer==2.1